+++ src/main/kotlin/net/lagerwey/plugins/cucumber/kotlin/CucumberKotlinLambdaExtension.kt
@@ -5,7 +5,6 @@ import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.progress.Task
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiManager
 import com.intellij.psi.SmartPointerManager
 import com.intellij.psi.search.GlobalSearchScope
@@ -14,10 +13,10 @@ import com.intellij.psi.search.PsiSearchHelper
 import com.intellij.psi.search.UsageSearchContext
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.util.indexing.FileBasedIndex
-import groovy.json.StringEscapeUtils
 import net.lagerwey.plugins.cucumber.kotlin.steps.KotlinLambdaStepDefinition
 import net.lagerwey.plugins.cucumber.kotlin.steps.KotlinParameterTypeManager
 import net.lagerwey.plugins.cucumber.kotlin.steps.KotlinStepDefinitionCreator
+import org.apache.commons.text.StringEscapeUtils
 import org.jetbrains.kotlin.idea.KotlinFileType
 import org.jetbrains.kotlin.idea.references.KtInvokeFunctionReference
 import org.jetbrains.kotlin.psi.KtCallExpression
@@ -32,7 +31,7 @@ class CucumberKotlinLambdaExtension : AbstractCucumberKotlinExtension() {
 
     override fun getStepDefinitionCreator(): StepDefinitionCreator = KotlinStepDefinitionCreator()
 
-    override fun loadStepsFor(featureFile: PsiFile?, module: Module): MutableList<AbstractStepDefinition> {
+    override fun loadStepsFor(module: Module): MutableList<AbstractStepDefinition> {
         val fileBasedIndex = FileBasedIndex.getInstance()
         val project = module.project